Donation page and organization of linked in files.
Put all CSS inside a styles folder and all JavaScript inside the
JavaScript folder and added pictures for each kind of card.

Created the general layout of the donation page and started creating a
custom object that will create the payment form for the donation on the
clicking of the next button.