Custom Objects and Methods/Properties for them
Created a method called createPayment() that will create the payment
form for the customer after they click the next button and then created
a custom object called details that has all of the inputs/labels for
the way the page comes originally and will reset the page.

After bug fixing I will make it so that the values the user puts in
carry over no matter which part of the form they click in/around.